+++ lib/src/pods/pod.dart
@@ -106,7 +106,8 @@ class Pod<T> extends BindWithMixinPodNotifier<T> {
 
   /// Casts [other] to `Pod<T>`.
   // ignore: invalid_use_of_visible_for_testing_member
-  static Pod<T> cast<T>(PodListenable<T> other) => other.asPodNotifier().asPod();
+  static Pod<T> cast<T>(PodListenable<T> other) =>
+      other.asPodNotifier().asPod();
 
   Pod.temp(super.value) : super.temp();