+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - chore: Remove this where unnecessary
 - chore: Restructure source files and improve docs
 - chore: Separate each exception into its own file
+- feat: Add .global and .temp constructors
 
 ## [0.2.3]